Fix a flaky test and reorganize the tests
