+++ mapreduce/tests/conftest.py
@@ -35,7 +35,6 @@ def dd_environment():
         compose_file=os.path.join(HERE, "compose", "docker-compose.yaml"),
         conditions=[WaitFor(setup_mapreduce, attempts=5, wait=5)],
         env_vars=env,
-        attempts=2,
     ):
         # 'custom_hosts' in metadata provides native /etc/hosts mappings in the agent's docker container
         yield INSTANCE_INTEGRATION, {'custom_hosts': get_custom_hosts()}
@@ -64,9 +63,7 @@ def mocked_auth_request():
 
 
 def get_custom_hosts():
-    # creat a mapping of mapreduce hostnames to localhost for DNS resolution
-    custom_hosts = [(host, '127.0.0.1') for host in MOCKED_E2E_HOSTS]
-    return custom_hosts
+    return [(host, '127.0.0.1') for host in MOCKED_E2E_HOSTS]
 
 
 def requests_get_mock(*args, **kwargs):
@@ -82,20 +79,19 @@ def requests_get_mock(*args, **kwargs):
                 "Apps URL must have the two query parameters: states=RUNNING and applicationTypes=MAPREDUCE"
             )
 
-    elif url == MR_JOBS_URL:
+    if url == MR_JOBS_URL:
         return MockResponse(file_path=os.path.join(HERE, "fixtures", "job_metrics"))
 
-    elif url == MR_JOB_COUNTERS_URL:
+    if url == MR_JOB_COUNTERS_URL:
         return MockResponse(file_path=os.path.join(HERE, "fixtures", "job_counter_metrics"))
 
-    elif url == MR_TASKS_URL:
+    if url == MR_TASKS_URL:
         return MockResponse(file_path=os.path.join(HERE, "fixtures", "task_metrics"))
 
-    elif url == CLUSTER_INFO_URL:
+    if url == CLUSTER_INFO_URL:
         return MockResponse(file_path=os.path.join(HERE, "fixtures", "cluster_info"))
 
-    else:
-        raise Exception("There is no mock request for {}".format(url))
+    raise Exception("There is no mock request for {}".format(url))
 
 
 def requests_auth_mock(*args, **kwargs):
